classes match

+++ c++_implementation/ui_code/circle_class.cpp
@@ -9,7 +9,7 @@ Circle::Circle(){
 
 }
 
-Circle::Circle(float r, float x, float y) {
+Circle::Circle(int r, int x, int y) {
 
     new_info = new Info;
